Move tests to __test__ folder.
