+++ src/redux/sagas/__tests__/authSaga.test.js
@@ -2,24 +2,24 @@ import { select } from 'redux-saga/effects';
 import { expectSaga, testSaga } from 'redux-saga-test-plan';
 import * as matchers from 'redux-saga-test-plan/matchers';
 
-import { authorize, authenticationFlow, loginSuccess } from './authSaga';
+import { authorize, authenticationFlow, loginSuccess } from '../authSaga';
 import {
   getAuthToken,
   LOGIN,
   LOGIN_SUCCESS,
   LOGIN_FAILURE,
   STORE_AUTH_TOKEN,
-} from '../reducers/auth';
-import * as API from './api';
-import * as Strings from '../../constants/strings';
+} from '../../reducers/auth';
+import * as API from '../api';
+import * as Strings from '../../../constants/strings';
 import {
   uiActions,
   getLoginScreenUsernameInputText,
   getLoginScreenPasswordInputText,
   SHOW_ERROR_ALERT,
-} from '../reducers/ui';
+} from '../../reducers/ui';
 
-import reducers from '../reducers';
+import reducers from '../../reducers';
 
 const storedToken = {
   access_token: 'c5f6b996-47aa-4c59-8fc7-8a03fcf5da9d',
@@ -96,7 +96,7 @@ it('runs authentication flow', () =>
     .withReducer(reducers)
     .provide([[matchers.call.fn(API.authorizeUser), { ok: true, data: storedToken }]])
     .dispatch({ type: LOGIN })
-    .delay(100)
+    .delay(1)
     .dispatch({ type: STORE_AUTH_TOKEN, token: null })
     .run(1000));
